Fixed the return value of getCurrentTime to return an Int32 on the managed side. This fixes the huge overflow that was being returned by this function.
git-svn-id: https://192.168.3.139:8443/svn/Engine/trunk@665 146ae76c-2b19-ef4c-9a69-67eb2b888ad7

+++ PCPlatformPlugin/Timer/PCSystemTimer.cs
@@ -42,6 +42,6 @@ namespace PCPlatform
         private static extern bool PerformanceCounter_initialize(IntPtr counter);
 
         [DllImport("PCPlatform")]
-        private static extern long PerformanceCounter_getCurrentTime(IntPtr counter);
+        private static extern Int32 PerformanceCounter_getCurrentTime(IntPtr counter);
     }
 }
